+++ D3D12Bundles/Shaders.cs
@@ -1,6 +1,6 @@
-﻿using D3D12HelloWorld.Rendering;
-using DirectX12GameEngine.Shaders;
+﻿using DirectX12GameEngine.Shaders;
 using System.Numerics;
+using wired.Rendering;
 
 namespace D3D12Bundles {
     public struct VSInput
