+++ D3D12HelloWorld/Graphics/PipelineState.cs
@@ -4,7 +4,7 @@ using System.Runtime.CompilerServices;
 using Vortice.Direct3D12;
 using Vortice.DXGI;
 
-namespace D3D12HelloWorld.Rendering {
+namespace wired.Graphics {
     public sealed class PipelineState : IDisposable {
         public ID3D12RootSignature RootSignature { get; }
