+++ D3D12HelloWorld/Graphics/Rendering.cs
@@ -1,173 +1,9 @@
-﻿using DirectX12GameEngine.Shaders;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Numerics;
+﻿using System;
 using System.Runtime.CompilerServices;
 using Vortice.Direct3D12;
 using Vortice.DXGI;
 
-namespace D3D12HelloWorld.Rendering {
-    [ShaderType("Texture2D")]
-    [ShaderResourceView]
-    public class Texture2D<T> : ShaderResourceView where T : unmanaged {
-        public Texture2D(ShaderResourceView shaderResourceView) : base(shaderResourceView) {
-        }
-
-        public T Sample(Sampler sampler, Vector2 textureCoordinate) => throw new NotImplementedException("This is intentionally not implemented.");
-    }
-
-    [ShaderType("Texture2D")]
-    [ShaderResourceView]
-    public class Texture2D : Texture2D<Vector4> {
-        public Texture2D(ShaderResourceView shaderResourceView) : base(shaderResourceView) {
-        }
-    }
-
-    public sealed class DescriptorAllocator : IDisposable {
-        private const int DescriptorsPerHeap = 4096;
-
-        readonly object mAllocatorLock = new object();
-
-        readonly DescriptorHeapDescription mDescription;
-
-        public int CurrentDescriptorCount { get; private set; }
-        //public int DescriptorCapacity { get; private set; }
-        public int DescriptorHandleIncrementSize { get; }
-        //public DescriptorHeapFlags Flags => mDescription.Flags;
-        internal ID3D12DescriptorHeap DescriptorHeap { get; }
-
-        public DescriptorAllocator(ID3D12Device device, DescriptorHeapType descriptorHeapType, int descriptorCount = DescriptorsPerHeap, DescriptorHeapFlags descriptorHeapFlags = DescriptorHeapFlags.None) {
-            if (descriptorCount < 1 || descriptorCount > DescriptorsPerHeap) {
-                throw new ArgumentOutOfRangeException(nameof(descriptorCount), $"Descriptor count must be between 1 and {DescriptorsPerHeap}.");
-            }
-
-            //Type = descriptorHeapType;
-            //Flags = descriptorHeapFlags;
-
-            DescriptorHandleIncrementSize = device.GetDescriptorHandleIncrementSize(descriptorHeapType);
-            mDescription = new DescriptorHeapDescription(descriptorHeapType, descriptorCount, descriptorHeapFlags);
-            DescriptorHeap = device.CreateDescriptorHeap(mDescription);
-            //DescriptorCapacity = descriptorCount;
-        }
-
-        public CpuDescriptorHandle Allocate(int count) {
-            lock (mAllocatorLock) {
-                if (count < 1 || count > mDescription.DescriptorCount) {
-                    throw new ArgumentOutOfRangeException(nameof(count), "Count must be between 1 and the total descriptor count.");
-                }
-
-                if (CurrentDescriptorCount + count > mDescription.DescriptorCount) {
-                    Reset();
-                }
-
-                CpuDescriptorHandle result = DescriptorHeap.GetCPUDescriptorHandleForHeapStart().Offset(CurrentDescriptorCount, DescriptorHandleIncrementSize);
-                CurrentDescriptorCount += count;
-                return result;
-            }
-        }
-
-        public CpuDescriptorHandle AllocateSlot(int slot) {
-            lock (mAllocatorLock) {
-                if (slot < 0 || slot > mDescription.DescriptorCount - 1) {
-                    throw new ArgumentOutOfRangeException(nameof(slot), "Slot must be between 0 and the total descriptor count - 1.");
-                }
-
-                CpuDescriptorHandle descriptor = DescriptorHeap.GetCPUDescriptorHandleForHeapStart().Offset(slot, DescriptorHandleIncrementSize);
-                return descriptor;
-            }
-        }
-
-        public GpuDescriptorHandle GetGpuDescriptorHandle(CpuDescriptorHandle descriptor) {
-            if (!mDescription.Flags.HasFlag(DescriptorHeapFlags.ShaderVisible)) {
-                throw new InvalidOperationException();
-            }
-
-            var offset = (int)(descriptor - DescriptorHeap.GetCPUDescriptorHandleForHeapStart());
-            return new GpuDescriptorHandle(DescriptorHeap.GetGPUDescriptorHandleForHeapStart(), offset);
-        }
-
-        public void Dispose() {
-            DescriptorHeap.Dispose();
-        }
-
-        public void Reset() {
-            CurrentDescriptorCount = 0;
-        }
-    }
-
-    public sealed class DescriptorSet {
-        readonly ID3D12Device mDevice;
-        public DescriptorHeapType DescriptorHeapType { get; }
-        public int CurrentDescriptorCount { get; private set; }
-        public int DescriptorCapacity { get; private set; }
-        public DescriptorAllocator DescriptorAllocator { get; }
-        public CpuDescriptorHandle StartCpuDescriptorHandle { get; }
-
-        public DescriptorSet(GraphicsDevice device, IEnumerable<Sampler> samplers)
-            : this(device, samplers.Count(), DescriptorHeapType.Sampler) {
-            AddSamplers(samplers);
-        }
-
-        public DescriptorSet(GraphicsDevice device, int descriptorCount, DescriptorHeapType descriptorHeapType = DescriptorHeapType.ConstantBufferViewShaderResourceViewUnorderedAccessView) {
-            if (descriptorCount < 1) {
-                throw new ArgumentOutOfRangeException("descriptorCount");
-            }
-
-            mDevice = device.NativeDevice;
-            DescriptorHeapType = descriptorHeapType;
-            DescriptorCapacity = descriptorCount;
-            DescriptorAllocator = GetDescriptorAllocator(device);
-            StartCpuDescriptorHandle = DescriptorAllocator.Allocate(DescriptorCapacity);
-        }
-
-        public void AddResourceViews(IEnumerable<ResourceView> resources) {
-            AddDescriptors(resources.Select((ResourceView r) => r.CpuDescriptorHandle));
-        }
-
-        public void AddResourceViews(params ResourceView[] resources) {
-            AddResourceViews(resources.AsEnumerable());
-        }
-
-        public void AddSamplers(IEnumerable<Sampler> samplers) {
-            AddDescriptors(samplers.Select((Sampler r) => r.CpuDescriptorHandle));
-        }
-
-        public void AddSamplers(params Sampler[] samplers) {
-            AddSamplers(samplers);
-        }
-
-        private void AddDescriptors(IEnumerable<CpuDescriptorHandle> descriptors) {
-            if (descriptors.Any()) {
-                CpuDescriptorHandle[] array = descriptors.ToArray();
-                if (CurrentDescriptorCount + array.Length > DescriptorCapacity) {
-                    throw new InvalidOperationException();
-                }
-
-                int[] array2 = new int[array.Length];
-                for (int i = 0; i < array2.Length; i++) {
-                    array2[i] = 1;
-                }
-
-                var value = new CpuDescriptorHandle(StartCpuDescriptorHandle, CurrentDescriptorCount, DescriptorAllocator.DescriptorHandleIncrementSize);
-                mDevice.CopyDescriptors(1, new CpuDescriptorHandle[1] { value, }, new int[1] { array.Length }, array.Length, array, array2, DescriptorHeapType);
-                CurrentDescriptorCount += array.Length;
-            }
-        }
-
-        DescriptorAllocator GetDescriptorAllocator(GraphicsDevice device) {
-            DescriptorHeapType descriptorHeapType = DescriptorHeapType;
-
-            DescriptorAllocator result = descriptorHeapType switch {
-                DescriptorHeapType.ConstantBufferViewShaderResourceViewUnorderedAccessView => device.ShaderResourceViewAllocator,
-                DescriptorHeapType.Sampler => device.SamplerAllocator,
-                _ => throw new NotSupportedException("This descriptor heap type is not supported."),
-            };
-
-            return result;
-        }
-    }
-
+namespace wired.Graphics {
     public abstract class ResourceView {
         protected ResourceView(GraphicsResource resource, CpuDescriptorHandle descriptor) {
             Resource = resource;
@@ -224,14 +60,6 @@ namespace D3D12HelloWorld.Rendering {
     public class DepthStencilView : ResourceView {
         public DepthStencilView(GraphicsResource resource, DescriptorAllocator dsvAllocator, DepthStencilViewDescription? description = null)
             : base(resource, CreateDepthStencilView(resource, dsvAllocator, description)) {
-        }
-
-        public static Texture CreateBuffer(GraphicsDevice device, int width, int height, Format format) {
-            var description = ResourceDescription.Texture2D(format, (uint)width, (uint)height, 1, 0, 1, 0, ResourceFlags.AllowDepthStencil);
-            var depthOptimizedClearValue = new ClearValue(format, 1.0f, 0);
-            var resource = device.NativeDevice.CreateCommittedResource(HeapProperties.DefaultHeapProperties, HeapFlags.None, description,
-                                                                       ResourceStates.DepthWrite, depthOptimizedClearValue);
-            return new Texture(device, resource);
         }
 
         public static DepthStencilView FromTexture2D(GraphicsResource resource, DescriptorAllocator dsvAllocator) {
@@ -375,51 +203,4 @@ namespace D3D12HelloWorld.Rendering {
         }
     }
 
-    public class MaterialPass {
-        public int PassIndex { get; set; }
-
-        public PipelineState? PipelineState { get; set; }
-
-        public DescriptorSet? ShaderResourceViewDescriptorSet { get; set; }
-
-        public DescriptorSet? SamplerDescriptorSet { get; set; }
-    }
-
-    //public class Material {
-    //    public MaterialDescriptor? Descriptor { get; set; }
-
-    //    public IList<MaterialPass> Passes { get; } = new List<MaterialPass>();
-
-    //    public static Task<Material> CreateAsync(GraphicsDevice device, MaterialDescriptor descriptor, IContentManager contentManager) {
-    //        Material material = new Material { Descriptor = descriptor };
-
-    //        MaterialGeneratorContext context = new MaterialGeneratorContext(device, material, contentManager);
-    //        return MaterialGenerator.GenerateAsync(descriptor, context);
-    //    }
-    //}
-
-    public sealed class MeshDraw {
-        public IndexBufferView? IndexBufferView { get; set; }
-
-        public VertexBufferView[]? VertexBufferViews { get; set; }
-    }
-
-    public sealed class Mesh {
-        public Mesh(MeshDraw meshDraw) {
-            MeshDraw = meshDraw;
-        }
-
-        public int MaterialIndex { get; set; }
-
-        public MeshDraw MeshDraw { get; set; }
-
-        public Matrix4x4 WorldMatrix { get; set; } = Matrix4x4.Identity;
-    }
-
-    public sealed class Model {
-        //public IList<Material> Materials { get; } = new List<Material>();
-        public IList<MaterialPass> Materials { get; } = new List<MaterialPass>();
-
-        public IList<Mesh> Meshes { get; } = new List<Mesh>();
-    }
 }
