+++ D3D12HelloWorld/Rendering/CompiledShader.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace D3D12HelloWorld.Rendering {
+namespace wired.Rendering {
     public class CompiledShader {
         public IDictionary<string, byte[]> Shaders { get; } = new Dictionary<string, byte[]>();
